Switch from YankRing to YankStack.
YankRing version 18.0 contains an annoying bug wherein the yanking
contents are messed up when pasting over a visual selection. The
contents of the visual selection are pushed into the yanking, causing
the next paste to insert the previous visual selection, instead of the
most recent yank, as expected. YankStack version 1.0.5 does not seem to
have this problem.